feat: 推广管理功能完成 + Docker 一键部署
- 完成推广统计后端 API (PromoStats 控制器)
- 修复管理员访问权限问题 (AdminAccessibleTrait)
- 修复缓存判断逻辑 (PromoStatsService)
- 新增 Docker 一键部署脚本 (setup.sh)
- 新增 Nginx 配置 + HTTPS 支持
- 修复前端路由和组件问题

+++ api/config/cache.php
@@ -35,7 +35,7 @@ return [
             // 密码
             'password' => env('redis.password', ''),
             // 缓存有效期 0表示永久缓存
-            'expire' => env('redis.expire', 0),
+            'expire' => (int) env('redis.expire', 0),
             // 缓存前缀
             'prefix' => env('redis.prefix', 'xmt:'),
             // 数据库
